Update DoubleTransformer.java
Joysticks very rarely center on `0`. Most of the time when you let go of a controller you end up with a value between `[-0.03, 0.03]`. In a lot of applications it's good practice to treat values less than a threshold as `0`.

Your change was incorrect, the existing code treats numbers less than the threshold as `0.05`, meaning that your robot will always command a spin and a drive.

It seems like you may have wanted to raise the default dead-zone threshold to `0.05`. I've done that here.